test(bundle-source): don't repeat yourself
