+++ packages/bundle-source/test/test-bigint-transform.js
@@ -4,6 +4,6 @@ import bundleSource from '..';
 
 test('bigint transform', async t => {
   const bundle = await bundleSource(`${__dirname}/../demo/bigint`, 'getExport');
-  console.log(bundle.source);
-  t.not(bundle.source.indexOf('37n'), -1);
+  // console.log(bundle.source);
+  t.assert(bundle.source.indexOf('37n') >= 0);
 });
